Fix npota-14: Error when checking what date the button "Afgelopen maand" gives
It does not give exactly one month ago, but 30 days ago. But when it's March on a leap year, it's 31 days ago.

I already put this in the script, but the fixture it used was a bit messy. It was maybe not so clear to someone, who changed it into exactly one month ago.